# 0-back/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9001",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get all settings",
        "operationId": "getAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SettingsDTO"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update a setting by name (only value is used)",
        "operationId": "updateOne",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/listes-deroulantes/{id}": {
      "get": {
        "tags": [
          "Listes déroulantes"
        ],
        "summary": "Get a liste déroulante by ID",
        "operationId": "getListeDeroulanteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListeDeroulanteDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Listes déroulantes"
        ],
        "summary": "Update a liste déroulante",
        "operationId": "updateListeDeroulante",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListeDeroulanteDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListeDeroulanteDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Listes déroulantes"
        ],
        "summary": "Delete a liste déroulante",
        "operationId": "deleteListeDeroulante",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/droits-libertes/{id}": {
      "get": {
        "tags": [
          "Droits \u0026 Libertés"
        ],
        "summary": "Get a droit/liberté by ID",
        "operationId": "getDroitLiberteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DroitLiberteDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Droits \u0026 Libertés"
        ],
        "summary": "Update a droit/liberté",
        "operationId": "updateDroitLiberte",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DroitLiberteDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DroitLiberteDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Droits \u0026 Libertés"
        ],
        "summary": "Delete a droit/liberté",
        "operationId": "deleteDroitLiberte",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/listes-deroulantes": {
      "post": {
        "tags": [
          "Listes déroulantes"
        ],
        "summary": "Create a new liste déroulante",
        "operationId": "createListeDeroulante",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListeDeroulanteDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListeDeroulanteDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/droits-libertes": {
      "post": {
        "tags": [
          "Droits \u0026 Libertés"
        ],
        "summary": "Create a new droit/liberté",
        "operationId": "createDroitLiberte",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DroitLiberteDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DroitLiberteDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions-qpc-cc/search": {
      "post": {
        "tags": [
          "Décisions QPC CC"
        ],
        "summary": "Search paginated décisions QPC CC",
        "operationId": "searchDecisions",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [
                "id.desc"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionQpcCcSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDTODecisionQpcCcDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/decisions-qpc-cc/{id}": {
      "get": {
        "tags": [
          "Décisions QPC CC"
        ],
        "summary": "Get a décision QPC CC by ID",
        "operationId": "getDecisionsById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionQpcCcDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SettingsDTO": {
        "required": [
          "categorie",
          "id",
          "nom",
          "nomEnum",
          "type",
          "updatedAt",
          "updatedById",
          "updatedByNom",
          "updatedByPrenom",
          "valeur"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "categorie": {
            "type": "string",
            "enum": [
              "GENERAL"
            ]
          },
          "nom": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "valeur": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "STRING",
              "NUMBER",
              "DATE",
              "BOOLEAN"
            ]
          },
          "nomEnum": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedById": {
            "type": "integer",
            "format": "int64"
          },
          "updatedByNom": {
            "type": "string"
          },
          "updatedByPrenom": {
            "type": "string"
          }
        }
      },
      "ListeDeroulanteDTO": {
        "required": [
          "actif",
          "champ",
          "id",
          "valeur"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "champ": {
            "type": "string"
          },
          "valeur": {
            "type": "string"
          },
          "actif": {
            "type": "boolean"
          }
        }
      },
      "DroitLiberteDTO": {
        "required": [
          "id",
          "texte"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "texte": {
            "type": "string"
          }
        }
      },
      "DecisionQpcCcSearchRequest": {
        "type": "object",
        "properties": {
          "origineQpcId": {
            "type": "integer",
            "format": "int64"
          },
          "qualiteDemandeurId": {
            "type": "integer",
            "format": "int64"
          },
          "typeDispositionLegislativeId": {
            "type": "integer",
            "format": "int64"
          },
          "matiereId": {
            "type": "integer",
            "format": "int64"
          },
          "dispositifDecisionCcId": {
            "type": "integer",
            "format": "int64"
          },
          "traitementEffetsPassesId": {
            "type": "integer",
            "format": "int64"
          },
          "oraliteId": {
            "type": "integer",
            "format": "int64"
          },
          "qualiteTiersInterventionId": {
            "type": "integer",
            "format": "int64"
          },
          "reserveIncompetenceConseilId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DecisionQpcCcDTO": {
        "required": [
          "applicationTheorieChangementCirconstances",
          "autresRemarques",
          "caractereNotableDecision",
          "createdAt",
          "dateAbrogationDifferee",
          "dateDecision",
          "delaiAvantAbrogationMois",
          "demandeRecusation",
          "deport",
          "dispositifDecisionCc",
          "dispositionsLegislativesContestees",
          "droitsLibertes",
          "id",
          "identiteAssociations",
          "identiteCollectivites",
          "identiteDemandeur",
          "identiteEntreprises",
          "identitePersonnesPhysiques",
          "identiteSyndicatsApOp",
          "matiere",
          "motifInconstitutionnalite",
          "nomMembreDeporteOuRecuse",
          "nombreAssociations",
          "nombreCollectivitesTerritoriales",
          "nombreDroitsLibertesInvoques",
          "nombreEntreprises",
          "nombreInterventionsAdmises",
          "nombreMembresSieges",
          "nombrePersonnesPhysiques",
          "nombreSyndicatsApOp",
          "numero",
          "oralite",
          "origineQpc",
          "priseEnCompteInterpretationJurisprudentielle",
          "qualiteDemandeur",
          "qualiteTiersIntervention",
          "referenceDecisionConseil",
          "referenceDecisionTransmission",
          "reserveIncompetenceConseil",
          "reserveOpportunite",
          "techniquesControle",
          "traitementEffetsPasses",
          "typeDispositionLegislative",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "referenceDecisionConseil": {
            "type": "string"
          },
          "numero": {
            "type": "string"
          },
          "dateDecision": {
            "type": "string",
            "format": "date"
          },
          "origineQpc": {
            "type": "string"
          },
          "referenceDecisionTransmission": {
            "type": "string"
          },
          "qualiteDemandeur": {
            "type": "string"
          },
          "identiteDemandeur": {
            "type": "string"
          },
          "dispositionsLegislativesContestees": {
            "type": "string"
          },
          "typeDispositionLegislative": {
            "type": "string"
          },
          "matiere": {
            "type": "string"
          },
          "dispositifDecisionCc": {
            "type": "string"
          },
          "dateAbrogationDifferee": {
            "type": "string",
            "format": "date"
          },
          "delaiAvantAbrogationMois": {
            "type": "integer",
            "format": "int32"
          },
          "traitementEffetsPasses": {
            "type": "string"
          },
          "nombreMembresSieges": {
            "type": "integer",
            "format": "int32"
          },
          "demandeRecusation": {
            "type": "integer",
            "format": "int32"
          },
          "deport": {
            "type": "integer",
            "format": "int32"
          },
          "nomMembreDeporteOuRecuse": {
            "type": "string"
          },
          "oralite": {
            "type": "string"
          },
          "applicationTheorieChangementCirconstances": {
            "type": "boolean"
          },
          "nombreDroitsLibertesInvoques": {
            "type": "integer",
            "format": "int32"
          },
          "nombreInterventionsAdmises": {
            "type": "integer",
            "format": "int32"
          },
          "qualiteTiersIntervention": {
            "type": "string"
          },
          "nombrePersonnesPhysiques": {
            "type": "integer",
            "format": "int32"
          },
          "identitePersonnesPhysiques": {
            "type": "string"
          },
          "nombreAssociations": {
            "type": "integer",
            "format": "int32"
          },
          "identiteAssociations": {
            "type": "string"
          },
          "nombreEntreprises": {
            "type": "integer",
            "format": "int32"
          },
          "identiteEntreprises": {
            "type": "string"
          },
          "nombreSyndicatsApOp": {
            "type": "integer",
            "format": "int32"
          },
          "identiteSyndicatsApOp": {
            "type": "string"
          },
          "nombreCollectivitesTerritoriales": {
            "type": "integer",
            "format": "int32"
          },
          "identiteCollectivites": {
            "type": "string"
          },
          "reserveOpportunite": {
            "type": "boolean"
          },
          "reserveIncompetenceConseil": {
            "type": "string"
          },
          "priseEnCompteInterpretationJurisprudentielle": {
            "type": "boolean"
          },
          "techniquesControle": {
            "type": "string"
          },
          "motifInconstitutionnalite": {
            "type": "string"
          },
          "autresRemarques": {
            "type": "string"
          },
          "caractereNotableDecision": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "droitsLibertes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DroitLiberteDTO"
            }
          }
        }
      },
      "PageDTODecisionQpcCcDTO": {
        "required": [
          "content",
          "page",
          "size",
          "totalElements"
        ],
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecisionQpcCcDTO"
            }
          }
        }
      },
...
    }
  }
}